feat: use Static layout
